requirements.txt
librosa
numpy
matplotlib
seaborn
scikit-learn
torch
torchvision
torchaudio
pandas
scipy
blas
lapack
sklearn.preprocessing
ipykernel
jupyterlab
libsndfile
dataclasses
torchsummary
pickle
-e .